+++ modules/src/core/ics02_client/events.rs
@@ -1,99 +1,13 @@
 //! Types for the IBC events emitted from Tendermint Websocket by the client module.
 
 use serde_derive::{Deserialize, Serialize};
-use tendermint_rpc::abci::tag::Tag;
-use tendermint_rpc::abci::Event as AbciEvent;
 
 use crate::core::ics02_client::client_type::ClientType;
-use crate::core::ics02_client::error::Error;
 use crate::core::ics02_client::header::AnyHeader;
 use crate::core::ics02_client::height::Height;
 use crate::core::ics24_host::identifier::ClientId;
-use crate::events::{IbcEvent, IbcEventType};
+use crate::events::IbcEvent;
 use crate::prelude::*;
-
-/// The content of the `key` field for the attribute containing the height.
-const HEIGHT_ATTRIBUTE_KEY: &str = "height";
-
-/// The content of the `key` field for the attribute containing the client identifier.
-const CLIENT_ID_ATTRIBUTE_KEY: &str = "client_id";
-
-/// The content of the `key` field for the attribute containing the client type.
-const CLIENT_TYPE_ATTRIBUTE_KEY: &str = "client_type";
-
-/// The content of the `key` field for the attribute containing the height.
-const CONSENSUS_HEIGHT_ATTRIBUTE_KEY: &str = "consensus_height";
-
-/// The content of the `key` field for the header in update client event.
-const HEADER_ATTRIBUTE_KEY: &str = "header";
-
-pub fn try_from_tx(event: &AbciEvent) -> Option<IbcEvent> {
-    match event.type_str.parse() {
-        Ok(IbcEventType::CreateClient) => extract_attributes_from_tx(event)
-            .map(CreateClient)
-            .map(IbcEvent::CreateClient)
-            .ok(),
-        Ok(IbcEventType::UpdateClient) => match extract_attributes_from_tx(event) {
-            Ok(attributes) => Some(IbcEvent::UpdateClient(UpdateClient {
-                common: attributes,
-                header: extract_header_from_tx(event).ok(),
-            })),
-            Err(_) => None,
-        },
-        Ok(IbcEventType::ClientMisbehaviour) => extract_attributes_from_tx(event)
-            .map(ClientMisbehaviour)
-            .map(IbcEvent::ClientMisbehaviour)
-            .ok(),
-        Ok(IbcEventType::UpgradeClient) => extract_attributes_from_tx(event)
-            .map(UpgradeClient)
-            .map(IbcEvent::UpgradeClient)
-            .ok(),
-        _ => None,
-    }
-}
-
-fn extract_attributes_from_tx(event: &AbciEvent) -> Result<Attributes, Error> {
-    let mut attr = Attributes::default();
-
-    for tag in &event.attributes {
-        let key = tag.key.as_ref();
-        let value = tag.value.as_ref();
-        match key {
-            HEIGHT_ATTRIBUTE_KEY => {
-                attr.height = value
-                    .parse()
-                    .map_err(|e| Error::invalid_string_as_height(value.to_string(), e))?
-            }
-            CLIENT_ID_ATTRIBUTE_KEY => {
-                attr.client_id = value.parse().map_err(Error::invalid_client_identifier)?
-            }
-            CLIENT_TYPE_ATTRIBUTE_KEY => {
-                attr.client_type = value
-                    .parse()
-                    .map_err(|_| Error::unknown_client_type(value.to_string()))?
-            }
-            CONSENSUS_HEIGHT_ATTRIBUTE_KEY => {
-                attr.consensus_height = value
-                    .parse()
-                    .map_err(|e| Error::invalid_string_as_height(value.to_string(), e))?
-            }
-            _ => {}
-        }
-    }
-
-    Ok(attr)
-}
-
-pub fn extract_header_from_tx(event: &AbciEvent) -> Result<AnyHeader, Error> {
-    for tag in &event.attributes {
-        let key = tag.key.as_ref();
-        let value = tag.value.as_ref();
-        if key == HEADER_ATTRIBUTE_KEY {
-            return AnyHeader::decode_from_string(value);
-        }
-    }
-    Err(Error::missing_raw_header())
-}
 
 /// NewBlock event signals the committing & execution of a new block.
 // TODO - find a better place for NewBlock
@@ -139,36 +53,6 @@ impl Default for Attributes {
     }
 }
 
-/// Convert attributes to Tendermint ABCI tags
-///
-/// # Note
-/// The parsing of `Key`s and `Value`s never fails, because the
-/// `FromStr` instance of `tendermint::abci::tag::{Key, Value}`
-/// is infallible, even if it is not represented in the error type.
-/// Once tendermint-rs improves the API of the `Key` and `Value` types,
-/// we will be able to remove the `.parse().unwrap()` calls.
-impl From<Attributes> for Vec<Tag> {
-    fn from(a: Attributes) -> Self {
-        let height = Tag {
-            key: HEIGHT_ATTRIBUTE_KEY.parse().unwrap(),
-            value: a.height.to_string().parse().unwrap(),
-        };
-        let client_id = Tag {
-            key: CLIENT_ID_ATTRIBUTE_KEY.parse().unwrap(),
-            value: a.client_id.to_string().parse().unwrap(),
-        };
-        let client_type = Tag {
-            key: CLIENT_TYPE_ATTRIBUTE_KEY.parse().unwrap(),
-            value: a.client_type.as_str().parse().unwrap(),
-        };
-        let consensus_height = Tag {
-            key: CONSENSUS_HEIGHT_ATTRIBUTE_KEY.parse().unwrap(),
-            value: a.height.to_string().parse().unwrap(),
-        };
-        vec![height, client_id, client_type, consensus_height]
-    }
-}
-
 impl core::fmt::Display for Attributes {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> Result<(), core::fmt::Error> {
         write!(
@@ -204,16 +88,6 @@ impl From<Attributes> for CreateClient {
 impl From<CreateClient> for IbcEvent {
     fn from(v: CreateClient) -> Self {
         IbcEvent::CreateClient(v)
-    }
-}
-
-impl From<CreateClient> for AbciEvent {
-    fn from(v: CreateClient) -> Self {
-        let attributes = Vec::<Tag>::from(v.0);
-        AbciEvent {
-            type_str: IbcEventType::CreateClient.as_str().to_string(),
-            attributes,
-        }
     }
 }
 
@@ -266,23 +140,6 @@ impl From<UpdateClient> for IbcEvent {
     }
 }
 
-impl From<UpdateClient> for AbciEvent {
-    fn from(v: UpdateClient) -> Self {
-        let mut attributes = Vec::<Tag>::from(v.common);
-        if let Some(h) = v.header {
-            let header = Tag {
-                key: HEADER_ATTRIBUTE_KEY.parse().unwrap(),
-                value: h.encode_to_string().parse().unwrap(),
-            };
-            attributes.push(header);
-        }
-        AbciEvent {
-            type_str: IbcEventType::UpdateClient.as_str().to_string(),
-            attributes,
-        }
-    }
-}
-
 impl core::fmt::Display for UpdateClient {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> Result<(), core::fmt::Error> {
         write!(f, "{}", self.common)
@@ -324,16 +181,6 @@ impl From<ClientMisbehaviour> for IbcEvent {
     }
 }
 
-impl From<ClientMisbehaviour> for AbciEvent {
-    fn from(v: ClientMisbehaviour) -> Self {
-        let attributes = Vec::<Tag>::from(v.0);
-        AbciEvent {
-            type_str: IbcEventType::ClientMisbehaviour.as_str().to_string(),
-            attributes,
-        }
-    }
-}
-
 /// Signals a recent upgrade of an on-chain client (IBC Client).
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Deserialize, Serialize)]
 pub struct UpgradeClient(pub Attributes);
@@ -353,60 +200,5 @@ impl UpgradeClient {
 impl From<Attributes> for UpgradeClient {
     fn from(attrs: Attributes) -> Self {
         UpgradeClient(attrs)
-    }
-}
-
-impl From<UpgradeClient> for AbciEvent {
-    fn from(v: UpgradeClient) -> Self {
-        let attributes = Vec::<Tag>::from(v.0);
-        AbciEvent {
-            type_str: IbcEventType::UpgradeClient.as_str().to_string(),
-            attributes,
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::core::ics02_client::header::Header;
-    use crate::mock::header::MockHeader;
-
-    #[test]
-    fn client_event_to_abci_event() {
-        let height = Height::new(1, 1);
-        let attributes = Attributes {
-            height,
-            client_id: "test_client".parse().unwrap(),
-            client_type: ClientType::Tendermint,
-            consensus_height: height,
-        };
-        let mut abci_events = vec![];
-        let create_client = CreateClient::from(attributes.clone());
-        abci_events.push(AbciEvent::from(create_client.clone()));
-        let client_misbehaviour = ClientMisbehaviour::from(attributes.clone());
-        abci_events.push(AbciEvent::from(client_misbehaviour.clone()));
-        let upgrade_client = UpgradeClient::from(attributes.clone());
-        abci_events.push(AbciEvent::from(upgrade_client.clone()));
-        let mut update_client = UpdateClient::from(attributes);
-        let header = MockHeader::new(height).wrap_any();
-        update_client.header = Some(header);
-        abci_events.push(AbciEvent::from(update_client.clone()));
-
-        for event in abci_events {
-            match try_from_tx(&event) {
-                Some(e) => match e {
-                    IbcEvent::CreateClient(e) => assert_eq!(e.0, create_client.0),
-                    IbcEvent::ClientMisbehaviour(e) => assert_eq!(e.0, client_misbehaviour.0),
-                    IbcEvent::UpgradeClient(e) => assert_eq!(e.0, upgrade_client.0),
-                    IbcEvent::UpdateClient(e) => {
-                        assert_eq!(e.common, update_client.common);
-                        assert_eq!(e.header, update_client.header);
-                    }
-                    _ => panic!("unexpected event type"),
-                },
-                None => panic!("converted event was wrong"),
-            }
-        }
     }
 }
