+++ relayer/src/chain/cosmos/query/tx.rs
@@ -1,10 +1,8 @@
 use ibc::core::ics02_client::client_consensus::QueryClientEventRequest;
-use ibc::core::ics02_client::events as ClientEvents;
 use ibc::core::ics04_channel::channel::QueryPacketEventDataRequest;
-use ibc::core::ics04_channel::events as ChannelEvents;
 use ibc::core::ics04_channel::packet::{Packet, Sequence};
 use ibc::core::ics24_host::identifier::ChainId;
-use ibc::events::{from_tx_response_event, IbcEvent};
+use ibc::events::IbcEvent;
 use ibc::query::{QueryTxHash, QueryTxRequest};
 use ibc::Height as ICSHeight;
 use tendermint_rpc::abci::transaction::Hash as TxHash;
@@ -15,6 +13,9 @@ use tendermint_rpc::{Client, HttpClient, Order, Url};
 
 use crate::chain::cosmos::query::{header_query, packet_query, tx_hash_query};
 use crate::error::Error;
+use crate::event::from_tx_response_event;
+use crate::event::ics02_client as client_events;
+use crate::event::ics04_channel as channel_events;
 
 /// This function queries transactions for events matching certain criteria.
 /// 1. Client Update request - returns a vector with at most one update client event
@@ -157,7 +158,7 @@ fn update_client_from_tx_search_response(
         .events
         .into_iter()
         .filter(|event| event.type_str == request.event_id.as_str())
-        .flat_map(|event| ClientEvents::try_from_tx(&event))
+        .flat_map(|event| client_events::try_from_tx(&event))
         .flat_map(|event| match event {
             IbcEvent::UpdateClient(mut update) => {
                 update.common.height = height;
@@ -218,7 +219,7 @@ fn filter_matching_event(
         return None;
     }
 
-    let ibc_event = ChannelEvents::try_from_tx(&event)?;
+    let ibc_event = channel_events::try_from_tx(&event)?;
     match ibc_event {
         IbcEvent::SendPacket(ref send_ev) if matches_packet(request, seq, &send_ev.packet) => {
             Some(ibc_event)
